Split Resolves into separate files
* Remove unused imports
